Unreviewed build fix after r213628.
Use 'size_t' for the function definition to match the header file. If you do not,
the 32-bit build fails..

* crypto/gnutls/CryptoAlgorithmECDHGnuTLS.cpp:
(WebCore::CryptoAlgorithmECDH::platformDeriveBits):



Canonical link: https://commits.webkit.org/186420@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213663 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/crypto/gnutls/CryptoAlgorithmECDHGnuTLS.cpp
@@ -32,7 +32,7 @@
 
 namespace WebCore {
 
-void CryptoAlgorithmECDH::platformDeriveBits(Ref<CryptoKey>&&, Ref<CryptoKey>&&, unsigned long, Callback&&, ScriptExecutionContext&, WorkQueue&)
+void CryptoAlgorithmECDH::platformDeriveBits(Ref<CryptoKey>&&, Ref<CryptoKey>&&, size_t, Callback&&, ScriptExecutionContext&, WorkQueue&)
 {
     notImplemented();
 }
